pyproject.toml
[project]
name = "dt-aid"
version = "0.1.0"
description = "AI-assisted auto-tagging for darktable libraries (faces + visual elements)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "darktable-ai-aid" }]

dependencies = [
    "insightface>=0.7.3",
    "onnxruntime>=1.17",
    "numpy>=1.26",
    "opencv-python-headless>=4.9",
    "pillow>=10.0",
    "rawpy>=0.21",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.4",
    "pyarrow>=15.0",
    "lxml>=5.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "psutil>=5.9",
    "rich>=13.7",
    "tqdm>=4.66",
    "platformdirs>=4.2",
]

[project.optional-dependencies]
elements = [
    "open-clip-torch>=2.24",
    "torch>=2.2",
]
yolo = [
    "ultralytics>=8.1",
]
gpu = [
    "onnxruntime-gpu>=1.24",
    "nvidia-cublas-cu12",
    "nvidia-cuda-runtime-cu12",
    "nvidia-cuda-nvrtc-cu12",
    "nvidia-cudnn-cu12",
    "nvidia-cufft-cu12",
    "nvidia-curand-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-cusparse-cu12",
    "nvidia-nccl-cu12",
    "nvidia-nvjitlink-cu12",
    "nvidia-nvtx-cu12",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]

[project.scripts]
dt-aid = "dt_aid.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dt_aid"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
